requirements.txt
asyncio==3.4.3
Pillow==9.5.0
pyaes==1.6.1
Pyrogram==2.0.103
PySocks==1.7.1
speedtest-cli==2.1.3
TgCrypto==1.2.5
